--- src/main/java/com/LetMeDoWith/LetMeDoWith/service/SocialProviderAuthFactoryService.java
@@ -1,21 +1,0 @@
-package com.LetMeDoWith.LetMeDoWith.service;
-
-import com.LetMeDoWith.LetMeDoWith.client.AuthClient;
-import com.LetMeDoWith.LetMeDoWith.enums.SocialProvider;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationContext;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class SocialProviderAuthFactoryService {
-	
-	private final ApplicationContext applicationContext;
-	
-	public AuthClient getClient(SocialProvider socialProvider) {
-		
-		return (AuthClient) applicationContext.getBean(
-			socialProvider.getCode().toLowerCase() + "AuthClient");
-		
-	}
-}
